Open to nested action test
